resolve migration

+++ database/migrations/2025_11_03_012827_desa.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('desa', function (Blueprint $table) {
             $table->id('iddesa')->primary();
-            $table->integer('idkecamatan')->nullable();
+            $table->unsignedBigInteger('idkecamatan')->nullable();
             $table->text('namadesa')->nullable();
 
             // Foreign key constraint
